Split out modules-specific declaration handling from SemaDecl.cpp into a new SemaModule.cpp.
git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@358633 91177308-0d34-0410-b5e6-96231b3b80d8
